+++ M2TWinForms.Demo/Form1.cs
@@ -11,7 +11,7 @@ using System.Windows.Forms;
 
 namespace M2TWinForms.Demo
 {
-    public partial class Form1 : M2TForm
+    public partial class Form1 : M2TFormBorderless
     {
         public Form1()
         {
