+++ M2TWinForms/Controls/Window/M2TFormBorderless.resx
